requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
click==8.2.1
fastapi==0.116.0
h11==0.16.0
hmmlearn==0.3.3
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
